fix norminette
