+++ incl/parsing/file_scene_parsing.h
@@ -1,0 +1,8 @@
+#ifndef FILE_SCENE_PARSING_H
+# define FILE_SCENE_PARSING_H
+
+# include "ft_raytracing.h"
+
+int	file_parsing(char *name, t_scene *scene);
+
+#endif
